Job Sphere Project Files Are Create

+++ src/main/java/com/jobportal/jobsphere/Repository/CompanyRepository.java
@@ -3,6 +3,4 @@ package com.jobportal.jobsphere.Repository;
 import com.jobportal.jobsphere.Entity.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface CompanyRepository extends JpaRepository<Company, BigInteger> {}
+public interface CompanyRepository extends JpaRepository<Company, Long> {}
