+++ src/main/java/com/jobportal/jobsphere/Repository/EducationRepository.java
@@ -3,6 +3,4 @@ package com.jobportal.jobsphere.Repository;
 import com.jobportal.jobsphere.Entity.Education;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface EducationRepository extends JpaRepository<Education, BigInteger> {}
+public interface EducationRepository extends JpaRepository<Education, Long> {}
