+++ src/main/java/com/jobportal/jobsphere/Repository/UserRepository.java
@@ -3,6 +3,4 @@ package com.jobportal.jobsphere.Repository;
 import com.jobportal.jobsphere.Entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface UserRepository extends JpaRepository<User,BigInteger> {}
+public interface UserRepository extends JpaRepository<User,Long> {}
