+++ src/main/java/com/jobportal/jobsphere/Repository/UserRolesRepository.java
@@ -3,6 +3,4 @@ package com.jobportal.jobsphere.Repository;
 import com.jobportal.jobsphere.Entity.User_Roles;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.math.BigInteger;
-
-public interface UserRolesRepository extends JpaRepository<User_Roles, BigInteger> {}
+public interface UserRolesRepository extends JpaRepository<User_Roles, Long> {}
